rename pal/system to pal/systemInfo
