+++ cc.config.json
@@ -56,7 +56,7 @@
             "overrides": {
                 "pal/minigame": "pal/minigame/non-minigame.ts",
                 "pal/audio": "pal/audio/web/player.ts",
-                "pal/system": "pal/system/web/system.ts",
+                "pal/systemInfo": "pal/system-info/web/system-info.ts",
                 "pal/input": "pal/input/web/index.ts"
             }
         },
@@ -66,7 +66,7 @@
             "overrides": {
                 "pal/minigame": "pal/minigame/non-minigame.ts",
                 "pal/audio": "pal/audio/native/player.ts",
-                "pal/system": "pal/system/native/system.ts",
+                "pal/systemInfo": "pal/system-info/native/system-info.ts",
                 "pal/input": "pal/input/native/index.ts"
             }
         },
@@ -76,7 +76,7 @@
             "overrides": {
                 "pal/minigame": "pal/minigame/{{context.platform.toLowerCase()}}.ts",
                 "pal/audio": "pal/audio/minigame/player.ts",
-                "pal/system": "pal/system/minigame/system.ts",
+                "pal/systemInfo": "pal/system-info/minigame/system-info.ts",
                 "pal/input": "pal/input/minigame/index.ts"
             }
         },
@@ -86,7 +86,7 @@
             "overrides": {
                 "pal/minigame": "pal/minigame/runtime.ts",
                 "pal/audio": "pal/audio/minigame/player.ts",
-                "pal/system": "pal/system/minigame/system.ts",
+                "pal/systemInfo": "pal/system-info/minigame/system-info.ts",
                 "pal/input": "pal/input/minigame/index.ts"
             }
         },
