+++ pal/audio/native/player.ts
@@ -1,10 +1,10 @@
-import { system } from 'pal/system';
+import { systemInfo } from 'pal/systemInfo';
 import { AudioType, AudioState, AudioEvent } from '../type';
 import { EventTarget } from '../../../cocos/core/event/event-target';
 import { legacyCC } from '../../../cocos/core/global-exports';
 import { clamp, clamp01 } from '../../../cocos/core';
 import { enqueueOperation, OperationInfo, OperationQueueable } from '../operation-queue';
-import { Platform } from '../../system/enum-type';
+import { Platform } from '../../system-info/enum-type';
 
 const urlCount: Record<string, number> = {};
 const audioEngine = jsb.AudioEngine;
@@ -113,7 +113,7 @@ export class AudioPlayer implements OperationQueueable {
     }
     static loadNative (url: string): Promise<unknown> {
         return new Promise((resolve, reject) => {
-            if (system.platform === Platform.WIN32) {
+            if (systemInfo.platform === Platform.WIN32) {
                 // NOTE: audioEngine.preload() not works well on Win32 platform.
                 // Especially when there is not audio output device.
                 resolve(url);
