+++ pal/input/web/accelerometer.ts
@@ -1,7 +1,7 @@
 import { AccelerometerCallback, AccelerometerInputEvent } from 'pal/input';
-import { system } from 'pal/system';
+import { systemInfo } from 'pal/systemInfo';
 import { EventTarget } from '../../../cocos/core/event/event-target';
-import { BrowserType, OS } from '../../system/enum-type';
+import { BrowserType, OS } from '../../system-info/enum-type';
 import { legacyCC } from '../../../cocos/core/global-exports';
 import { SystemEventType } from '../../../cocos/core/platform/event-manager/event-enum';
 
@@ -21,7 +21,7 @@ export class AccelerometerInputSource {
         // init event name
         this._globalEventClass = window.DeviceMotionEvent || window.DeviceOrientationEvent;
         // TODO fix DeviceMotionEvent bug on QQ Browser version 4.1 and below.
-        if (system.browserType === BrowserType.MOBILE_QQ) {
+        if (systemInfo.browserType === BrowserType.MOBILE_QQ) {
             this._globalEventClass = window.DeviceOrientationEvent;
         }
         this._deviceEventName = this._globalEventClass === window.DeviceMotionEvent ? 'devicemotion' : 'deviceorientation';
@@ -87,8 +87,8 @@ export class AccelerometerInputSource {
         }
 
         // fix android acc values are opposite
-        if (system.os === OS.ANDROID
-            && system.browserType !== BrowserType.MOBILE_QQ) {
+        if (systemInfo.os === OS.ANDROID
+            && systemInfo.browserType !== BrowserType.MOBILE_QQ) {
             x = -x;
             y = -y;
         }
