+++ pal/system-info/enum-type/browser-type.ts
