+++ pal/system-info/enum-type/event.ts
