+++ pal/system-info/enum-type/index.ts
