+++ pal/system-info/enum-type/network-type.ts
