+++ pal/system-info/enum-type/operating-system.ts
