+++ pal/system-info/enum-type/platform.ts
