+++ pal/system-info/minigame/system-info.ts
@@ -1,33 +1,38 @@
-import { SafeAreaEdge, SupportCapability } from 'pal/system';
+import { ALIPAY, BAIDU, BYTEDANCE, COCOSPLAY, HUAWEI, LINKSURE, OPPO, QTT, VIVO, WECHAT, XIAOMI, DEBUG, EDITOR, TEST } from 'internal:constants';
+import { SafeAreaEdge, SupportCapability } from 'pal/systemInfo';
+import { minigame } from 'pal/minigame';
 import { Size } from '../../../cocos/core/math';
 import { EventTarget } from '../../../cocos/core/event/event-target';
 import { BrowserType, NetworkType, Orientation, OS, Platform, AppEvent, Language } from '../enum-type';
 
-// these value is defined in the native layer
-const orientationMap: Record<string, Orientation> = {
-    0: Orientation.PORTRAIT,
-    '-90': Orientation.LANDSCAPE_LEFT,
-    90: Orientation.LANDSCAPE_RIGHT,
-    180: Orientation.PORTRAIT_UPSIDE_DOWN,
-};
-const networkTypeMap: Record<string, NetworkType> = {
-    0: NetworkType.NONE,
-    1: NetworkType.LAN,
-    2: NetworkType.WWAN,
-};
-const platformMap: Record<number, Platform> = {
-    0: Platform.WIN32,
-    // 1 is Linux platform in native engine
-    2: Platform.MACOS,
-    3: Platform.ANDROID,
-    // 4 is IPHONE
-    4: Platform.IOS,
-    // 5 is IPAD
-    5: Platform.IOS,
-    6: Platform.OHOS,
-};
+// NOTE: register minigame platform here
+let currentPlatform: Platform;
+if (WECHAT) {
+    currentPlatform = Platform.WECHAT_GAME;
+} else if (BAIDU) {
+    currentPlatform = Platform.BAIDU_MINI_GAME;
+} else if (XIAOMI) {
+    currentPlatform = Platform.XIAOMI_QUICK_GAME;
+} else if (ALIPAY) {
+    currentPlatform = Platform.ALIPAY_MINI_GAME;
+} else if (BYTEDANCE) {
+    currentPlatform = Platform.BYTEDANCE_MINI_GAME;
+} else if (OPPO) {
+    currentPlatform = Platform.OPPO_MINI_GAME;
+} else if (VIVO) {
+    currentPlatform = Platform.VIVO_MINI_GAME;
+} else if (HUAWEI) {
+    currentPlatform = Platform.HUAWEI_QUICK_GAME;
+} else if (COCOSPLAY) {
+    currentPlatform = Platform.COCOSPLAY;
+} else if (LINKSURE) {
+    currentPlatform = Platform.LINKSURE_MINI_GAME;
+} else if (QTT) {
+    currentPlatform = Platform.QTT_MINI_GAME;
+}
 
 class System {
+    public readonly networkType: NetworkType;
     public readonly isNative: boolean;
     public readonly isBrowser: boolean;
     public readonly isMobile: boolean;
@@ -42,7 +47,6 @@ class System {
     public readonly browserVersion: string;
     public readonly pixelRatio: number;
     public readonly supportCapability: SupportCapability;
-    // TODO: need to wrap the function __isObjectValid()
 
     public get isOnFullScreen (): boolean {
         return false;
@@ -50,17 +54,11 @@ class System {
 
     private _eventTarget: EventTarget = new EventTarget();
 
-    public get networkType (): NetworkType {
-        return networkTypeMap[jsb.device.getNetworkType()];
-    }
-
     constructor () {
-        this.isNative = true;
+        const minigameSysInfo = minigame.getSystemInfoSync();
+        this.networkType = NetworkType.LAN;  // TODO
+        this.isNative = false;
         this.isBrowser = false;
-
-        // @ts-expect-error __getPlatform()
-        this.platform = platformMap[__getPlatform()];
-        this.isMobile = this.platform === Platform.ANDROID || this.platform === Platform.IOS || this.platform === Platform.OHOS;
 
         // init isLittleEndian
         this.isLittleEndian = (() => {
@@ -71,27 +69,49 @@ class System {
         })();
 
         // init languageCode and language
-        // @ts-expect-error __getCurrentLanguageCode() defined in JSB
-        const currLanguage = __getCurrentLanguageCode();
-        this.nativeLanguage = currLanguage ? currLanguage.toLowerCase() : Language.UNKNOWN;
-        // @ts-expect-error __getCurrentLanguage() defined in JSB
-        this.language = __getCurrentLanguage();
+        this.nativeLanguage = minigameSysInfo.language;
+        this.language = minigameSysInfo.language.substr(0, 2) as Language;
 
-        // @ts-expect-error __getOS() defined in JSB
-        this.os = __getOS();
-        // @ts-expect-error __getOSVersion() defined in JSB
-        this.osVersion = __getOSVersion();
+        // init os, osVersion and osMainVersion
+        const minigamePlatform = minigameSysInfo.platform.toLocaleLowerCase();
+        if (minigamePlatform === 'android') {
+            this.os = OS.ANDROID;
+        } else if (minigamePlatform === 'ios') {
+            this.os = OS.IOS;
+        } else if (minigamePlatform === 'windows') {
+            this.os = OS.WINDOWS;
+        } else {
+            this.os = OS.UNKNOWN;
+        }
+        let minigameSystem = minigameSysInfo.system.toLowerCase();
+        // Adaptation to Android P
+        if (minigameSystem === 'android p') {
+            minigameSystem = 'android p 9.0';
+        }
+        const version = /[\d.]+/.exec(minigameSystem);
+        this.osVersion = version ? version[0] : minigameSystem;
         this.osMainVersion = parseInt(this.osVersion);
+
+        // init isMobile and platform
+        this.platform = currentPlatform;
+        this.isMobile = this.os !== OS.WINDOWS;
 
         // init browserType and browserVersion
         this.browserType = BrowserType.UNKNOWN;
         this.browserVersion = '';
 
-        this.pixelRatio = jsb.device.getDevicePixelRatio() || 1;
+        this.pixelRatio = minigameSysInfo.pixelRatio;
 
         // init capability
+        const _tmpCanvas1 = document.createElement('canvas');  // TODO: remove this
+        let supportWebp;
+        try {
+            supportWebp = TEST ? false : _tmpCanvas1.toDataURL('image/webp').startsWith('data:image/webp');
+        } catch (e) {
+            supportWebp  = false;
+        }
         this.supportCapability = {
-            webp: true,
+            webp: supportWebp,  // TODO
             gl: true,
             canvas: true,
             imageBitmap: false,
@@ -102,50 +122,38 @@ class System {
     }
 
     private _registerEvent () {
-        jsb.onResize = (size) => {
-            if (size.width === 0 || size.height === 0) return;
-            size.width /= this.pixelRatio;
-            size.height /= this.pixelRatio;
-
-            // TODO: remove this function calling
-            window.resize(size.width, size.height);
-            this._eventTarget.emit(AppEvent.RESIZE);
-        };
-        jsb.onOrientationChanged = (event) => {
-            this._eventTarget.emit(AppEvent.ORIENTATION_CHANGE);
-        };
-        jsb.onPause = () => {
+        // TODO: onResize or onOrientationChange is not supported well
+        minigame.onHide(() => {
             this._eventTarget.emit(AppEvent.HIDE);
-        };
-        jsb.onResume = () => {
+        });
+        minigame.onShow(() => {
             this._eventTarget.emit(AppEvent.SHOW);
-        };
-        jsb.onClose = () => {
-            this._eventTarget.emit(AppEvent.CLOSE);
-        };
+        });
     }
 
     public resizeScreen (size: Size): Promise<void> {
-        return Promise.reject(new Error('screen resize has not been supported yet on this platform.'));
+        return Promise.reject(new Error('screen resize is not supported on this platform.'));
     }
     public requestFullScreen (): Promise<void> {
-        return Promise.reject(new Error('request fullscreen has not been supported yet on this platform.'));
+        return Promise.reject(new Error('request fullscreen is not supported on this platform.'));
     }
     public exitFullScreen (): Promise<void> {
-        return Promise.reject(new Error('exit fullscreen has not been supported yet on this platform.'));
+        return Promise.reject(new Error('exit fullscreen is not supported on this platform.'));
     }
     public getScreenSize (): Size {
-        return new Size(window.innerWidth, window.innerHeight);
+        const sysInfo = minigame.getSystemInfoSync();
+        return new Size(sysInfo.screenWidth, sysInfo.screenHeight);
     }
     public getOrientation (): Orientation {
-        return orientationMap[jsb.device.getDeviceOrientation()];
+        return minigame.orientation;
     }
     public getSafeAreaEdge (): SafeAreaEdge {
-        const nativeSafeArea = jsb.device.getSafeAreaEdge();
-        let topEdge = nativeSafeArea.x;
-        let bottomEdge = nativeSafeArea.z;
-        let leftEdge = nativeSafeArea.y;
-        let rightEdge = nativeSafeArea.w;
+        const minigameSafeArea = minigame.getSafeArea();
+        const screenSize = this.getScreenSize();
+        let topEdge = minigameSafeArea.top;
+        let bottomEdge = screenSize.height - minigameSafeArea.bottom;
+        let leftEdge = minigameSafeArea.left;
+        let rightEdge = screenSize.width - minigameSafeArea.right;
         const orientation = this.getOrientation();
         // Make it symmetrical.
         if (orientation === Orientation.PORTRAIT) {
@@ -167,13 +175,15 @@ class System {
         };
     }
     public getBatteryLevel (): number {
-        return jsb.device.getBatteryLevel();
+        return minigame.getBatteryInfoSync().level / 100;
     }
     public triggerGC (): void {
-        jsb.garbageCollect();
+        minigame.triggerGC();
     }
     public openURL (url: string): void {
-        jsb.openURL(url);
+        if (DEBUG) {
+            console.warn('openURL is not supported');
+        }
     }
     public now (): number {
         if (Date.now) {
@@ -183,13 +193,13 @@ class System {
         return +(new Date());
     }
     public restartJSVM (): void {
-        // @ts-expect-error __restartVM() is defined in JSB
-        __restartVM();
+        if (DEBUG) {
+            console.warn('restartJSVM is not supported.');
+        }
     }
 
     public close () {
-        // @ts-expect-error __close() is defined in JSB
-        __close();
+        // TODO: minigame.exitMiniProgram() not implemented.
     }
 
     public onHide (cb: () => void) {
