+++ pal/system-info/web/screen-utils.ts
